Moving view-related stuff in its own class
