+++ src/mainwindow.cpp
@@ -23,23 +23,8 @@
 #include <QNetworkReply>
 #include <QNetworkRequest>
 #include <QJsonDocument>
-#include "QVTKOpenGLNativeWidget.h"
-#include "vtkGenericOpenGLRenderWindow.h"
-#include "vtkRenderer.h"
-#include "vtkAxesActor.h"
-#include "vtkOrientationMarkerWidget.h"
 #include "vtkExtractBlock.h"
-#include "vtkProperty.h"
 #include "vtkCamera.h"
-#include "vtkCubeAxesActor.h"
-#include "vtkUnstructuredGrid.h"
-#include "vtkCompositeDataGeometryFilter.h"
-#include "vtkCaptionActor2D.h"
-#include "vtkTextProperty.h"
-#include "vtkMapper.h"
-#include "vtkCellData.h"
-#include "vtkDoubleArray.h"
-#include "vtkProperty2D.h"
 #include "infowindow.h"
 #include "aboutdlg.h"
 #include "licensedlg.h"
@@ -56,21 +41,13 @@
 #include "vtkreader.h"
 #include "stlreader.h"
 #include "colorprofile.h"
-#include "ointeractorstyle2d.h"
-#include "ointeractorstyle3d.h"
 #include "selection.h"
+#include "view.h"
 #include "common/loadfileevent.h"
 #include "common/notificationwidget.h"
 #include "common/infowidget.h"
 
 static const int MAX_RECENT_FILES = 10;
-
-QColor MainWindow::SIDESET_CLR = QColor(255, 173, 79);
-QColor MainWindow::SIDESET_EDGE_CLR = QColor(26, 26, 102);
-QColor MainWindow::NODESET_CLR = QColor(168, 91, 2);
-
-float MainWindow::EDGE_WIDTH = 1;
-float MainWindow::OUTLINE_WIDTH = 1.5;
 
 // Main window - Load thread
 
@@ -130,21 +107,11 @@ MainWindow::MainWindow(QWidget * parent) :
     file_watcher(new QFileSystemWatcher()),
     notification(nullptr),
     file_changed_notification(nullptr),
-    render_mode(SHADED_WITH_EDGES),
     color_profile_idx(0),
     menu_bar(new QMenuBar(nullptr)),
     recent_menu(nullptr),
     export_menu(nullptr),
-    view_menu(nullptr),
-    view_mode(nullptr),
     view(nullptr),
-    render_window(nullptr),
-    renderer(nullptr),
-    interactor(nullptr),
-    ori_marker(nullptr),
-    cube_axes_actor(nullptr),
-    interactor_style_2d(nullptr),
-    interactor_style_3d(nullptr),
     info_dock(nullptr),
     info_window(nullptr),
     about_dlg(nullptr),
@@ -157,29 +124,20 @@ MainWindow::MainWindow(QWidget * parent) :
     open_action(nullptr),
     close_action(nullptr),
     clear_recent_file(nullptr),
-    shaded_action(nullptr),
-    shaded_w_edges_action(nullptr),
-    hidden_edges_removed_action(nullptr),
-    transluent_action(nullptr),
     view_info_wnd_action(nullptr),
+    tools_explode_action(nullptr),
     tools_mesh_quality_action(nullptr),
-    perspective_action(nullptr),
-    ori_marker_action(nullptr),
     minimize(nullptr),
     bring_all_to_front(nullptr),
     show_main_window(nullptr),
     about_box_action(nullptr),
     check_update_action(nullptr),
     view_license_action(nullptr),
-    visual_repr(nullptr),
     color_profile_action_group(nullptr),
     windows_action_group(nullptr),
     namgr(new QNetworkAccessManager())
 {
     connect(this->namgr, &QNetworkAccessManager::finished, this, &MainWindow::onHttpReply);
-
-    this->interactor_style_2d = new OInteractorStyle2D(this);
-    this->interactor_style_3d = new OInteractorStyle3D(this);
 
     QSize default_size = QSize(1000, 700);
     QVariant geom = this->settings->value("window/geometry", default_size);
@@ -196,12 +154,9 @@ MainWindow::MainWindow(QWidget * parent) :
     setAcceptDrops(true);
 
     connectSignals();
-    setupVtk();
+    this->view->setupVtk();
     this->mesh_quality_tool->setupVtk();
     setColorProfile();
-
-    setupOrientationMarker();
-    setupCubeAxesActor();
 
     clear();
     show();
@@ -217,8 +172,6 @@ MainWindow::~MainWindow()
     delete this->file_watcher;
     delete this->menu_bar;
     delete this->view;
-    this->render_window->Delete();
-    this->renderer->Delete();
     delete this->info_window;
     delete this->info_dock;
     delete this->explode_tool;
@@ -226,8 +179,6 @@ MainWindow::~MainWindow()
     for (auto & it : this->color_profiles)
         delete it;
     delete this->load_thread;
-    delete this->view_menu;
-    delete this->view_mode;
     delete this->notification;
     delete this->file_changed_notification;
     delete this->windows_action_group;
@@ -261,41 +212,29 @@ MainWindow::getCenterOfBounds() const
     return this->center_of_bounds;
 }
 
-vtkRenderer *
-MainWindow::getRenderer() const
-{
-    return this->renderer;
-}
-
-vtkGenericOpenGLRenderWindow *
-MainWindow::getRenderWindow() const
-{
-    return this->render_window;
-}
-
 QSettings *
 MainWindow::getSettings()
 {
     return this->settings;
 }
 
-QWidget *
+View *
 MainWindow::getView()
 {
     return this->view;
 }
 
+const BlockObject *
+MainWindow::getSelectedBlock()
+{
+    return this->select_tool->getSelectedBlock();
+}
+
 void
 MainWindow::setupWidgets()
 {
-    this->view = new QVTKOpenGLNativeWidget();
+    this->view = new View(this);
     setCentralWidget(this->view);
-
-    this->render_window = vtkGenericOpenGLRenderWindow::New();
-    this->view->setRenderWindow(this->render_window);
-
-    this->renderer = vtkRenderer::New();
-    this->render_window->AddRenderer(this->renderer);
 
     this->info_window = new InfoWindow();
 
@@ -306,7 +245,7 @@ MainWindow::setupWidgets()
     this->info_dock->setWidget(this->info_window);
     addDockWidget(Qt::RightDockWidgetArea, this->info_dock);
 
-    setupViewModeWidget(this);
+    this->view->setupWidgets();
     setupFileChangedNotificationWidget();
     setupNotificationWidget();
     this->select_tool->setupWidgets();
@@ -318,60 +257,6 @@ MainWindow::setupWidgets()
 void
 MainWindow::setupViewModeWidget(QMainWindow * wnd)
 {
-    this->view_menu = new QMenu();
-    this->shaded_action = this->view_menu->addAction("Shaded");
-    this->shaded_action->setCheckable(true);
-    this->shaded_action->setShortcut(QKeySequence("Ctrl+1"));
-    this->shaded_w_edges_action = this->view_menu->addAction("Shaded with edges");
-    this->shaded_w_edges_action->setCheckable(true);
-    this->shaded_w_edges_action->setShortcut(QKeySequence("Ctrl+2"));
-    this->hidden_edges_removed_action = this->view_menu->addAction("Hidden edges removed");
-    this->hidden_edges_removed_action->setCheckable(true);
-    this->hidden_edges_removed_action->setShortcut(QKeySequence("Ctrl+3"));
-    this->transluent_action = this->view_menu->addAction("Transluent");
-    this->transluent_action->setCheckable(true);
-    this->transluent_action->setShortcut(QKeySequence("Ctrl+4"));
-    this->shaded_w_edges_action->setChecked(true);
-    this->render_mode = SHADED_WITH_EDGES;
-
-    this->visual_repr = new QActionGroup(this->view_menu);
-    this->visual_repr->addAction(this->shaded_action);
-    this->visual_repr->addAction(this->shaded_w_edges_action);
-    this->visual_repr->addAction(this->hidden_edges_removed_action);
-    this->visual_repr->addAction(this->transluent_action);
-    this->visual_repr->setExclusive(true);
-
-    this->view_menu->addSeparator();
-    this->perspective_action = this->view_menu->addAction("Perspective");
-    this->perspective_action->setCheckable(true);
-    this->perspective_action->setChecked(true);
-
-    this->view_menu->addSeparator();
-    this->ori_marker_action = this->view_menu->addAction("Orientation marker");
-    this->ori_marker_action->setCheckable(true);
-    this->ori_marker_action->setChecked(true);
-
-    connect(this->shaded_action, &QAction::triggered, this, &MainWindow::onShadedTriggered);
-    connect(this->shaded_w_edges_action,
-            &QAction::triggered,
-            this,
-            &MainWindow::onShadedWithEdgesTriggered);
-    connect(this->hidden_edges_removed_action,
-            &QAction::triggered,
-            this,
-            &MainWindow::onHiddenEdgesRemovedTriggered);
-    connect(this->transluent_action, &QAction::triggered, this, &MainWindow::onTransluentTriggered);
-    connect(this->perspective_action, &QAction::toggled, this, &MainWindow::onPerspectiveToggled);
-    connect(this->ori_marker_action,
-            &QAction::toggled,
-            this,
-            &MainWindow::onOrientationMarkerVisibilityChanged);
-
-    this->view_mode = new QPushButton(wnd);
-    this->view_mode->setFixedSize(60, 32);
-    this->view_mode->setIcon(QPixmap(":/resources/render-mode.svg"));
-    this->view_mode->setMenu(this->view_menu);
-    this->view_mode->show();
 }
 
 void
@@ -426,10 +311,7 @@ MainWindow::setupMenuBar()
     file_menu->addAction("Quit", this, &QCoreApplication::quit, QKeySequence("Ctrl+Q"));
 
     QMenu * view_menu = this->menu_bar->addMenu("View");
-    view_menu->addAction(this->shaded_action);
-    view_menu->addAction(this->shaded_w_edges_action);
-    view_menu->addAction(this->hidden_edges_removed_action);
-    view_menu->addAction(this->transluent_action);
+    this->view->setMenu(view_menu);
     view_menu->addSeparator();
     this->view_info_wnd_action =
         view_menu->addAction("Info window", this, &MainWindow::onViewInfoWindow);
@@ -495,7 +377,7 @@ MainWindow::updateMenuBar()
     this->close_action->setEnabled(has_file);
 
     bool showing_mesh_quality = this->mesh_quality_tool->isVisible();
-    this->visual_repr->setEnabled(!showing_mesh_quality);
+    this->view->updateMenuBar(!showing_mesh_quality);
 }
 
 void
@@ -560,22 +442,10 @@ MainWindow::connectSignals()
 }
 
 void
-MainWindow::setupVtk()
-{
-    this->interactor = this->render_window->GetInteractor();
-
-    // TODO: set background from preferences/templates
-    this->renderer->SetGradientBackground(true);
-    // set anti-aliasing on
-    this->renderer->SetUseFXAA(true);
-    this->render_window->SetMultiSamples(1);
-}
-
-void
 MainWindow::clear()
 {
     this->mesh_quality_tool->done();
-    this->renderer->RemoveAllViewProps();
+    this->view->clear();
 
     for (auto & it : this->blocks)
         delete it.second;
@@ -627,53 +497,6 @@ MainWindow::hasFile()
 }
 
 void
-MainWindow::setupOrientationMarker()
-{
-    std::array<QColor, 3> clr({ QColor(188, 39, 26), QColor(65, 147, 41), QColor(0, 0, 200) });
-
-    vtkAxesActor * axes = vtkAxesActor::New();
-    axes->SetNormalizedTipLength(0, 0, 0);
-
-    std::array<vtkProperty *, 3> shaft_property({ axes->GetXAxisShaftProperty(),
-                                                  axes->GetYAxisShaftProperty(),
-                                                  axes->GetZAxisShaftProperty() });
-    for (std::size_t i = 0; i < shaft_property.size(); i++) {
-        auto prop = shaft_property[i];
-        prop->SetLineWidth(HIDPI(1));
-        prop->SetColor(clr[i].redF(), clr[i].greenF(), clr[i].blueF());
-    }
-
-    std::array<vtkCaptionActor2D *, 3> caption_actors({ axes->GetXAxisCaptionActor2D(),
-                                                        axes->GetYAxisCaptionActor2D(),
-                                                        axes->GetZAxisCaptionActor2D() });
-    for (std::size_t i = 0; i < caption_actors.size(); i++) {
-        auto text_prop = caption_actors[i]->GetCaptionTextProperty();
-        text_prop->SetColor(clr[i].redF(), clr[i].greenF(), clr[i].blueF());
-        text_prop->BoldOn();
-        text_prop->ItalicOff();
-        text_prop->ShadowOff();
-    }
-
-    this->ori_marker = vtkOrientationMarkerWidget::New();
-    this->ori_marker->SetDefaultRenderer(this->renderer);
-    this->ori_marker->SetOrientationMarker(axes);
-    this->ori_marker->SetViewport(0.8, 0, 1.0, 0.2);
-    this->ori_marker->SetInteractor(this->interactor);
-    this->ori_marker->SetEnabled(true);
-    this->ori_marker->SetInteractive(false);
-}
-
-void
-MainWindow::setupCubeAxesActor()
-{
-    this->cube_axes_actor = vtkCubeAxesActor::New();
-    this->cube_axes_actor->VisibilityOff();
-    this->cube_axes_actor->SetCamera(this->renderer->GetActiveCamera());
-    this->cube_axes_actor->SetGridLineLocation(vtkCubeAxesActor::VTK_GRID_LINES_ALL);
-    this->cube_axes_actor->SetFlyMode(vtkCubeAxesActor::VTK_FLY_OUTER_EDGES);
-}
-
-void
 MainWindow::computeTotalBoundingBox()
 {
     vtkBoundingBox bbox;
@@ -684,8 +507,7 @@ MainWindow::computeTotalBoundingBox()
 
     double bounds[6];
     bbox.GetBounds(bounds);
-    this->cube_axes_actor->SetBounds(bounds);
-    this->renderer->AddViewProp(this->cube_axes_actor);
+    this->view->setTotalBoundingBox(bounds);
     this->info_window->setBounds(bounds[0], bounds[1], bounds[2], bounds[3], bounds[4], bounds[5]);
 
     double center[3];
@@ -718,7 +540,7 @@ MainWindow::checkFileExists(const QString & file_name)
 void
 MainWindow::addBlocks()
 {
-    auto * camera = this->renderer->GetActiveCamera();
+    auto * camera = this->view->getActiveCamera();
     auto * reader = this->load_thread->getReader();
 
     for (auto & binfo : reader->getBlocks()) {
@@ -734,12 +556,8 @@ MainWindow::addBlocks()
         }
         else
             block = new BlockObject(reader->getVtkOutputPort(), camera);
-        setBlockProperties(block);
         this->blocks[binfo.number] = block;
-
-        this->renderer->AddViewProp(block->getActor());
-        this->renderer->AddViewProp(block->getSilhouetteActor());
-
+        this->view->addBlock(block);
         emit blockAdded(binfo.number, QString::fromStdString(binfo.name));
     }
 }
@@ -758,9 +576,7 @@ MainWindow::addSideSets()
 
         auto sideset = new SideSetObject(eb->GetOutputPort());
         this->side_sets[finfo.number] = sideset;
-        setSideSetProperties(sideset);
-        this->renderer->AddViewProp(sideset->getActor());
-
+        this->view->addSideSet(sideset);
         emit sideSetAdded(finfo.number, QString::fromStdString(finfo.name));
     }
 }
@@ -779,9 +595,7 @@ MainWindow::addNodeSets()
 
         auto * nodeset = new NodeSetObject(eb->GetOutputPort());
         this->node_sets[ninfo.number] = nodeset;
-        this->setNodeSetProperties(nodeset);
-        this->renderer->AddViewProp(nodeset->getActor());
-
+        this->view->addNodeSet(nodeset);
         emit nodeSetAdded(ninfo.number, QString::fromStdString(ninfo.name));
     }
 }
@@ -814,171 +628,6 @@ MainWindow::getNodeSet(int nodeset_id)
         return it->second;
     else
         return nullptr;
-}
-
-void
-MainWindow::setSelectedBlockProperties(BlockObject * block, bool highlighted)
-{
-    auto * property = block->getProperty();
-    if (this->render_mode == SHADED) {
-        property->SetColor(SIDESET_CLR.redF(), SIDESET_CLR.greenF(), SIDESET_CLR.blueF());
-        property->SetOpacity(1.0);
-        property->SetEdgeVisibility(false);
-    }
-    else if (this->render_mode == SHADED_WITH_EDGES) {
-        property->SetColor(SIDESET_CLR.redF(), SIDESET_CLR.greenF(), SIDESET_CLR.blueF());
-        property->SetOpacity(1.0);
-        property->SetEdgeVisibility(true);
-        property->SetEdgeColor(SIDESET_EDGE_CLR.redF(),
-                               SIDESET_EDGE_CLR.greenF(),
-                               SIDESET_EDGE_CLR.blueF());
-        property->SetLineWidth(HIDPI(EDGE_WIDTH));
-    }
-    else if (this->render_mode == HIDDEN_EDGES_REMOVED) {
-        property->SetColor(SIDESET_CLR.redF(), SIDESET_CLR.greenF(), SIDESET_CLR.blueF());
-        property->SetOpacity(1.0);
-        property->SetEdgeVisibility(false);
-    }
-    else if (this->render_mode == TRANSLUENT) {
-        property->SetColor(SIDESET_CLR.redF(), SIDESET_CLR.greenF(), SIDESET_CLR.blueF());
-        property->SetOpacity(0.33);
-        property->SetEdgeVisibility(false);
-    }
-
-    setHighlightedBlockProperties(block, highlighted);
-}
-
-void
-MainWindow::setDeselectedBlockProperties(BlockObject * block, bool highlighted)
-{
-    auto * property = block->getProperty();
-    if (this->render_mode == SHADED) {
-        auto & clr = block->getColor();
-        property->SetColor(clr.redF(), clr.greenF(), clr.blueF());
-        property->SetOpacity(block->getOpacity());
-        property->SetEdgeVisibility(false);
-    }
-    else if (this->render_mode == SHADED_WITH_EDGES) {
-        auto & clr = block->getColor();
-        property->SetColor(clr.redF(), clr.greenF(), clr.blueF());
-        property->SetOpacity(block->getOpacity());
-        property->SetEdgeVisibility(true);
-        property->SetEdgeColor(SIDESET_EDGE_CLR.redF(),
-                               SIDESET_EDGE_CLR.greenF(),
-                               SIDESET_EDGE_CLR.blueF());
-        property->SetLineWidth(HIDPI(EDGE_WIDTH));
-    }
-    else if (this->render_mode == HIDDEN_EDGES_REMOVED) {
-        property->SetColor(1., 1., 1.);
-        property->SetOpacity(1.0);
-        property->SetEdgeVisibility(false);
-    }
-    else if (this->render_mode == TRANSLUENT) {
-        auto & clr = block->getColor();
-        property->SetColor(clr.redF(), clr.greenF(), clr.blueF());
-        property->SetOpacity(0.33);
-        property->SetEdgeVisibility(false);
-    }
-
-    setHighlightedBlockProperties(block, highlighted);
-}
-
-void
-MainWindow::setHighlightedBlockProperties(BlockObject * block, bool highlighted)
-{
-    auto * property = block->getSilhouetteProperty();
-    if (highlighted) {
-        block->setSilhouetteVisible(true);
-
-        if (this->render_mode == SHADED) {
-            property->SetColor(SelectTool::HIGHLIGHT_CLR.redF(),
-                               SelectTool::HIGHLIGHT_CLR.greenF(),
-                               SelectTool::HIGHLIGHT_CLR.blueF());
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-        else if (this->render_mode == SHADED_WITH_EDGES) {
-            property->SetColor(SelectTool::HIGHLIGHT_CLR.redF(),
-                               SelectTool::HIGHLIGHT_CLR.greenF(),
-                               SelectTool::HIGHLIGHT_CLR.blueF());
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-        else if (this->render_mode == HIDDEN_EDGES_REMOVED) {
-            property->SetColor(SelectTool::HIGHLIGHT_CLR.redF(),
-                               SelectTool::HIGHLIGHT_CLR.greenF(),
-                               SelectTool::HIGHLIGHT_CLR.blueF());
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-        else if (this->render_mode == TRANSLUENT) {
-            property->SetColor(SelectTool::HIGHLIGHT_CLR.redF(),
-                               SelectTool::HIGHLIGHT_CLR.greenF(),
-                               SelectTool::HIGHLIGHT_CLR.blueF());
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-    }
-    else {
-        if (this->render_mode == SHADED) {
-            block->setSilhouetteVisible(false);
-        }
-        else if (this->render_mode == SHADED_WITH_EDGES) {
-            block->setSilhouetteVisible(false);
-        }
-        else if (this->render_mode == HIDDEN_EDGES_REMOVED) {
-            block->setSilhouetteVisible(true);
-            property->SetColor(0, 0, 0);
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-        else if (this->render_mode == TRANSLUENT) {
-            block->setSilhouetteVisible(true);
-            property->SetColor(0, 0, 0);
-            property->SetLineWidth(HIDPI(OUTLINE_WIDTH));
-        }
-    }
-}
-
-void
-MainWindow::setBlockProperties(BlockObject * block, bool selected, bool highlighted)
-{
-    auto * property = block->getProperty();
-    property->SetRepresentationToSurface();
-    property->SetAmbient(0.4);
-    property->SetDiffuse(0.6);
-    if (selected)
-        this->setSelectedBlockProperties(block, highlighted);
-    else
-        this->setDeselectedBlockProperties(block, highlighted);
-}
-
-void
-MainWindow::setSideSetProperties(SideSetObject * sideset)
-{
-    auto * property = sideset->getProperty();
-    property->SetColor(SIDESET_CLR.redF(), SIDESET_CLR.greenF(), SIDESET_CLR.blueF());
-    property->SetEdgeVisibility(false);
-    property->LightingOff();
-    if (this->render_mode == SHADED || this->render_mode == SHADED_WITH_EDGES) {
-        property->SetEdgeColor(SIDESET_EDGE_CLR.redF(),
-                               SIDESET_EDGE_CLR.greenF(),
-                               SIDESET_EDGE_CLR.blueF());
-        if (this->render_mode == SHADED_WITH_EDGES) {
-            property->SetLineWidth(HIDPI(EDGE_WIDTH));
-            property->SetEdgeVisibility(true);
-        }
-    }
-}
-
-void
-MainWindow::setNodeSetProperties(NodeSetObject * nodeset)
-{
-    auto * property = nodeset->getProperty();
-    property->SetRepresentationToPoints();
-    property->SetRenderPointsAsSpheres(true);
-    property->SetVertexVisibility(true);
-    property->SetEdgeVisibility(false);
-    property->SetPointSize(HIDPI(5));
-    property->SetColor(NODESET_CLR.redF(), NODESET_CLR.greenF(), NODESET_CLR.blueF());
-    property->SetOpacity(1);
-    property->SetAmbient(1);
-    property->SetDiffuse(0);
 }
 
 void
@@ -1033,12 +682,7 @@ MainWindow::setColorProfile()
     else
         profile = this->color_profiles[0];
 
-    const auto & qclr = profile->getColor("bkgnd");
-    double bkgnd[3] = { qclr.redF(), qclr.greenF(), qclr.blueF() };
-
-    this->renderer->SetBackground(bkgnd);
-    this->renderer->SetBackground2(bkgnd);
-
+    this->view->setColorProfile(profile);
     this->mesh_quality_tool->setColorProfile(profile);
 }
 
@@ -1222,90 +866,44 @@ MainWindow::loadIntoVtk()
     this->select_tool->loadIntoVtk(reader->getVtkOutputPort());
     this->mesh_quality_tool->loadIntoVtk();
 
-    if (reader->getDimensionality() == 3)
-        this->interactor->SetInteractorStyle(this->interactor_style_3d);
-    else
-        this->interactor->SetInteractorStyle(this->interactor_style_2d);
-
-    auto * camera = this->renderer->GetActiveCamera();
-    auto * focal_point = camera->GetFocalPoint();
-    camera->SetPosition(focal_point[0], focal_point[1], 1);
-    camera->SetRoll(0);
-    this->renderer->ResetCamera();
+    this->view->setInteractorStyle(reader->getDimensionality());
+    this->view->resetCamera();
 }
 
 void
 MainWindow::onBlockVisibilityChanged(int block_id, bool visible)
 {
-    BlockObject * block = getBlock(block_id);
-    if (block) {
-        block->setVisible(visible);
-        if (this->render_mode == HIDDEN_EDGES_REMOVED || this->render_mode == TRANSLUENT)
-            block->setSilhouetteVisible(visible);
-        else
-            block->setSilhouetteVisible(false);
-    }
+    this->view->setBlockVisibility(block_id, visible);
 }
 
 void
 MainWindow::onBlockOpacityChanged(int block_id, double opacity)
 {
-    BlockObject * block = getBlock(block_id);
-    if (block) {
-        block->setOpacity(opacity);
-        if (this->render_mode == SHADED || this->render_mode == SHADED_WITH_EDGES) {
-            auto * property = block->getProperty();
-            property->SetOpacity(opacity);
-        }
-    }
+    this->view->setBlockOpacity(block_id, opacity);
 }
 
 void
 MainWindow::onBlockColorChanged(int block_id, QColor color)
 {
-    BlockObject * block = getBlock(block_id);
-    if (block) {
-        block->setColor(color);
-        auto * property = block->getProperty();
-        if (this->render_mode == HIDDEN_EDGES_REMOVED)
-            property->SetColor(1, 1, 1);
-        else
-            property->SetColor(color.redF(), color.greenF(), color.blueF());
-    }
+    this->view->setBlockColor(block_id, color);
 }
 
 void
 MainWindow::onSideSetVisibilityChanged(int sideset_id, bool visible)
 {
-    auto * sideset = getSideSet(sideset_id);
-    if (sideset)
-        sideset->setVisible(visible);
+    this->view->setSideSetVisibility(sideset_id, visible);
 }
 
 void
 MainWindow::onNodeSetVisibilityChanged(int nodeset_id, bool visible)
 {
-    auto * nodeset = getNodeSet(nodeset_id);
-    if (nodeset)
-        nodeset->setVisible(visible);
+    this->view->setNodeSetVisibility(nodeset_id, visible);
 }
 
 void
 MainWindow::onCubeAxisVisibilityChanged(bool visible)
 {
-    if (visible)
-        this->cube_axes_actor->VisibilityOn();
-    else
-        this->cube_axes_actor->VisibilityOff();
-}
-
-void
-MainWindow::onOrientationMarkerVisibilityChanged(bool visible)
-{
-    if (visible)
-        this->ori_marker->EnabledOn();
-    else
-        this->ori_marker->EnabledOff();
+    this->view->setCubeAxisVisibility(visible);
 }
 
 void
@@ -1349,86 +947,9 @@ MainWindow::onNewFile()
 }
 
 void
-MainWindow::onShadedTriggered(bool checked)
-{
-    this->render_mode = SHADED;
-    for (auto & it : this->blocks) {
-        auto * block = it.second;
-        bool selected = this->select_tool->getSelectedBlock() == block;
-        setBlockProperties(block, selected);
-        block->setSilhouetteVisible(false);
-    }
-    for (auto & it : this->side_sets) {
-        auto * sideset = it.second;
-        setSideSetProperties(sideset);
-    }
-}
-
-void
-MainWindow::onShadedWithEdgesTriggered(bool checked)
-{
-    this->render_mode = SHADED_WITH_EDGES;
-    for (auto & it : this->blocks) {
-        auto * block = it.second;
-        bool selected = this->select_tool->getSelectedBlock() == block;
-        setBlockProperties(block, selected);
-        block->setSilhouetteVisible(false);
-    }
-    for (auto & it : this->side_sets) {
-        auto * sideset = it.second;
-        setSideSetProperties(sideset);
-    }
-}
-
-void
-MainWindow::onHiddenEdgesRemovedTriggered(bool checked)
-{
-    this->render_mode = HIDDEN_EDGES_REMOVED;
-    for (auto & it : this->blocks) {
-        auto * block = it.second;
-        bool selected = this->select_tool->getSelectedBlock() == block;
-        setBlockProperties(block, selected);
-        block->setSilhouetteVisible(block->visible());
-    }
-    for (auto & it : this->side_sets) {
-        auto * sideset = it.second;
-        setSideSetProperties(sideset);
-    }
-}
-
-void
-MainWindow::onTransluentTriggered(bool checked)
-{
-    this->render_mode = TRANSLUENT;
-    for (auto & it : this->blocks) {
-        auto * block = it.second;
-        bool selected = this->select_tool->getSelectedBlock() == block;
-        setBlockProperties(block, selected);
-        block->setSilhouetteVisible(block->visible());
-    }
-    for (auto & it : this->side_sets) {
-        auto * sideset = it.second;
-        setSideSetProperties(sideset);
-    }
-}
-
-void
-MainWindow::onPerspectiveToggled(bool checked)
-{
-    if (checked) {
-        auto * camera = this->renderer->GetActiveCamera();
-        camera->ParallelProjectionOff();
-    }
-    else {
-        auto * camera = this->renderer->GetActiveCamera();
-        camera->ParallelProjectionOn();
-    }
-}
-
-void
 MainWindow::onUpdateWindow()
 {
-    this->render_window->Render();
+    this->view->render();
 }
 
 void
@@ -1466,7 +987,7 @@ MainWindow::onViewInfoWindow()
     else
         this->info_dock->show();
     this->updateMenuBar();
-    this->updateViewModeLocation();
+    updateViewModeLocation();
 }
 
 void
@@ -1480,8 +1001,7 @@ MainWindow::onColorProfileTriggered(QAction * action)
 void
 MainWindow::updateViewModeLocation()
 {
-    auto tr = this->view->geometry().topRight();
-    this->view_mode->move(tr.x() - 10 - this->view_mode->width(), tr.y() + 10);
+    this->view->updateLocation();
 }
 
 void
@@ -1569,18 +1089,5 @@ MainWindow::getVersionFromReply(QNetworkReply * reply)
 void
 MainWindow::activateRenderMode()
 {
-    switch (this->render_mode) {
-    case SHADED:
-        onShadedTriggered(true);
-        break;
-    case SHADED_WITH_EDGES:
-        onShadedWithEdgesTriggered(true);
-        break;
-    case HIDDEN_EDGES_REMOVED:
-        onHiddenEdgesRemovedTriggered(true);
-        break;
-    case TRANSLUENT:
-        onTransluentTriggered(true);
-        break;
-    }
+    this->view->activateRenderMode();
 }
